feat(git): support cloning through archives in Github
Adds a new `CloneOptions` struct that can be passed to `Clone`. This
struct contains a `UseArchive` field, which defaults to `false`. When
set to true, an archive will be used instead of a git clone.

This should be much faster than a clone for larger repositories.
However, it has the drawback of no `.git` directory being present, hence
it not being on by default.

This is currently only supported on Github, Gitlab support may be added
in the future.