feat(token): cache Fetch tokens by default (#30)
Opts to cache tokens returned by `Fetch` by default. Adds a new
`Options` type that can be passed to `Fetch` to configure (disable) this
behaviour. As part of this, the `allowUnauthenticated` function option
has been moved to that struct. While the original bool is kept to
prevent breaking changes, it will be removed before v1 so callers should
update their calls ASAP.