fix(resolver): expose failure to resolve repository better
Previously, when listing remotes we'd only return the exit code, which
isn't very helpful. Now, using our internal `execerr` package, we also
return stderr in the error to aid in debugging.

Also included is a new err (`ErrUnableToSatisfy`) that makes it easier
to detect when the resolver failed to satisfy contraints vs reach the
ability to solve the constraints to begin with.