fix(git): force `LC_ALL=C` when calling the `git` CLI (#91)
## Problem

If a user sets their `LC_ALL` to a non-English value, and that language has a translation file for `git`, `git.DefaultBranch` will fail since it's [searching using the English output](https://github.com/jaredallard/vcs/blob/aec6089bfd68839472b6bba467b203c92e572400/git/git.go#L51).

## Solution

When invoking `git`, always set `LC_ALL=C`, which forces `git` to default to a non-translated version.

## Alternatives considered

Port to `go-git` - too much of a lift right now. If Jared wishes to do so in the future, there will be a version of this in `getoutreach/stencil` to copy from.

## Notes for reviewers

I set `go` to `1.24.0` so I could get `testing.T#Context()`. Let me know if that's not desired, and I'll revert it.